AppiumAutomation/requirements.txt
Appium-Python-Client==1.1.0
behave==1.2.6
requests==2.25.1
selenium==3.141.0